src/cloud/server/requirements.txt
mysql-connector-python==8.4.0
paho-mqtt==1.6.1
